\name{isoph-package}
\alias{isoph-package}
\docType{package}
\title{Isotonic Proportional Hazards Model}
\description{Nonparametric estimation of an isotonic covariate effect for proportional hazards model.}
\details{
\tabular{ll}{
Package: \tab isoph\cr
Type: \tab Package\cr
Version: \tab 1.1.3\cr
Date: \tab 2018-11-24\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Yunro Chung [cre]

Maintainer: Yunro Chung <yunro.chung@asu.edu>
}
\references{Yunro Chung, Anastasia Ivanova, Michael M. Hudgens, Jason P. Fine (2018), Partial likelihood estimation of isotonic proportional hazards models, Biometrika, 105(1), 133-148.}
\keyword{Isotonic regression, Survival analysis, Constrained estimation}
