\name{isoph-package}
\alias{isoph-package}
\docType{package}
\title{Isotonic Proportional Hazards Model}
\description{Nonparametric estimation of an isotonic covariate effect for Cox's partial likelihood.}
\details{
\tabular{ll}{
Package: \tab isoph\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2015-11-16\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Yunro Chung, Anastasia Ivanova, Michael M. Hudgens, Jason P. Fine

Maintainer: Yunro Chung <yunro@live.unc.edu>
}
\references{Yunro Chung, Anastasia Ivanova, Michael M. Hudgens, Jason P. Fine, Partial likelihood estimation of isotonic proportional hazards models (under revision).}
\keyword{Isotonic regression, Survival analysis, Constrained estimation}
