\name{isoph-package}
\alias{isoph-package}
\docType{package}
\title{Isotonic Proportional Hazards Model}
\description{Nonparametric estimation of an isotonic covariate effect for proportional hazards model.}
\details{
\tabular{ll}{
Package: \tab isoph\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2016-01-01\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Yunro Chung [cre], Anastasia Ivanova, Michael G. Hudgens and Jason P. Fine

Maintainer: Yunro Chung <yunro@live.unc.edu>
}
\references{Yunro Chung, Anastasia Ivanova, Michael M. Hudgens, Jason P. Fine, Partial likelihood estimation of isotonic proportional hazards models (under revision).}
\keyword{Isotonic regression, Survival analysis, Constrained estimation}
